# swappo-backend/docs/swagger.yaml
...
openapi: 3.0.0
info:
  title: Swappo Trading API
  description: Complete API documentation for the Swappo non-monetary trading platform
  version: 1.0.0
  contact:
    name: Swappo Team
    email: support@swappo.com
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://your-domain.com/api
    description: Production server
security:
  - BetterAuth: []
paths:
...
  /auth/session:
    get:
      tags: [Authentication]
      summary: Get current session
      security:
        - BetterAuth: []
      responses:
        200:
          description: Session information
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        401:
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Logout current user
      security:
        - BetterAuth: []
      responses:
        200:
          description: Logged out successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'

  # User 
...
  /users/{userId}/items:
    get:
      tags: [Users]
      summary: Get user's items
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [AVAILABLE, RESERVED, SWAPPED, REMOVED, ALL]
            default: AVAILABLE
      responses:
        200:
          description: List of user's items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
...
  /items:
    get:
      tags: [Items]
      summary: Get all available items (marketplace)
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: condition
          in: query
          schema:
            type: string
            enum: [NEW, GOOD, FAIR, POOR]
        - name: latitude
          in: query
          schema:
            type: number
            format: float
            minimum: -90
            maximum: 90
        - name: longitude
          in: query
          schema:
            type: number
            format: float
            minimum: -180
            maximum: 180
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: exclude_user
          in: query
          schema:
            type: string
      responses:
        200:
          description: Paginated list of items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Item'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      pages:
                        type: integer

    post:
      tags: [Items]
      summary: Create new item
      security:
        - BetterAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [title, description, category, condition]
              properties:
                title:
                  type: string
                  minLength: 3
                description:
                  type: string
                  minLength: 10
                category:
                  type: string
                condition:
                  type: string
                  enum: [NEW, GOOD, FAIR, POOR]
                latitude:
                  type: number
                  format: float
                  minimum: -90
                  maximum: 90
                longitude:
                  type: number
                  format: float
                  minimum: -180
                  maximum: 180
                images:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 5
      responses:
        201:
          description: Item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
  /items/{itemId}:
    get:
      tags: [Items]
      summary: Get item details with trade requests
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Item details with trade requests
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Item'
                  - type: object
                    properties:
                      trade_requests_for:
                        type: array
                        items:
                          $ref: '#/components/schemas/TradeRequest'
        404:
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Items]
      summary: Update item
      security:
        - BetterAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                category:
                  type: string
                condition:
                  type: string
                  enum: [NEW, GOOD, FAIR, POOR]
                latitude:
                  type: number
                  format: float
                  minimum: -90
                  maximum: 90
                longitude:
                  type: number
                  format: float
                  minimum: -180
                  maximum: 180
                removeImageIds:
                  type: string
                  description: JSON array of image IDs to remove
                newImages:
                  type: array
                  items:
                    type: string
                    format: binary


      responses:
        200:
          description: Item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        404:
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Items]
      summary: Delete item
      security:
        - BetterAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Item deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        400:
          description: Cannot delete item with pending trades
        404:
          $ref: '#/components/responses/NotFound'
...
  /items/{itemId}/status:
    patch:
      tags: [Items]
      summary: Toggle item status
      security:
        - BetterAuth: []
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [AVAILABLE, REMOVED]
      responses:
        200:
          description: Status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'



  # Trade
...
  /ratings:
    post:
      tags: [Ratings]
      summary: Create rating/review
      security:
        - BetterAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trade_id, reviewee_id, rating]
              properties:
                trade_id:
                  type: string
                reviewee_id:
                  type: string
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                comment:
                  type: string
      responses:
        201:
          description: Rating created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rating'
        400:
          $ref: '#/components/responses/BadRequest'
...
  /ratings/{ratingId}:
    put:
      tags: [Ratings]
      summary: Update rating (within 24 hours)
      security:
        - BetterAuth: []
      parameters:
        - name: ratingId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                comment:
                  type: string
      responses:
        200:
          description: Rating updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rating'
        400:
          description: Rating can only be edited within 24 hours
components:
  securitySchemes:
    BetterAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: BetterAuth session cookie
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: "clu1234567890"
        username:
          type: string
          example: "john_trader"
        email:
          type: string
          format: email
          example: "john@example.com"
        profile_picture:
          type: string
          nullable: true
          example: "/uploads/profiles/avatar.jpg"
        bio:
          type: string
          nullable: true
          example: "Love trading electronics and books!"
        loyalty_points:
          type: integer
          example: 150
        badge:
          type: string
          enum: [BRONZE, SILVER, GOLD, DIAMOND, RUBY]
          example: "GOLD"
        date_joined:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time
          nullable: true
    Item:
      type: object
      properties:
        id:
          type: string
          example: "cli1234567890"
        user_id:
          type: string
          example: "clu1234567890"
        title:
          type: string
          example: "iPhone 12 - Excellent Condition"
        description:
          type: string
          example: "Barely used iPhone 12, comes with original box and charger"
        category:
          type: string
          example: "Electronics"
        condition:
          type: string
          enum: [NEW, GOOD, FAIR, POOR]
          example: "GOOD"
        status:
          type: string
          enum: [AVAILABLE, RESERVED, SWAPPED, REMOVED]
          example: "AVAILABLE"
        latitude:
          type: number
          format: float
          nullable: true
          minimum: -90
          maximum: 90
          example: 6.9271
        longitude:
          type: number
          format: float
          nullable: true
          minimum: -180
          maximum: 180
          example: 79.8612
        posted_at:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
        images:
          type: array
          items:
            $ref: '#/components/schemas/ItemImage'
        _count:
          type: object
          properties:
            trade_requests_for:
              type: integer
    ItemImage:
      type: object
      properties:
        id:
          type: string
          example: "img1234567890"
        item_id:
          type: string
          example: "cli1234567890"
        url:
          type: string
          example: "/uploads/items/image.jpg"
    TradeRequest:
      type: object
      properties:
        id:
          type: string
          example: "tr1234567890"
        requested_item_id:
          type: string
          example: "cli1234567890"
        offered_item_id:
          type: string
          example: "cli0987654321"
        requester_id:
          type: string
          example: "clu1234567890"
        status:
          type: string
          enum: [PENDING, ACCEPTED, REJECTED, CANCELLED]
          example: "PENDING"
        requested_at:
          type: string
          format: date-time
        requested_item:
          $ref: '#/components/schemas/Item'
        offered_item:
          $ref: '#/components/schemas/Item'
        requester:
          $ref: '#/components/schemas/User'
...
    Rating:
      type: object
      properties:
        id:
          type: string
          example: "rat1234567890"
        trade_id:
          type: string
          example: "trd1234567890"
        reviewer_id:
          type: string
          example: "clu1234567890"
        reviewee_id:
          type: string
          example: "clu0987654321"
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          nullable: true
          example: "Great trader, very reliable!"
        created_at:
          type: string
          format: date-time
        reviewer:
          $ref: '#/components/schemas/User'
        reviewee:
          $ref: '#/components/schemas/User'
        trade:
          type: object
          properties:
            requested_item:
              type: object
              properties:
                title:
                  type: string
            offered_item:
              type: object
              properties:
                title:
                  type: string
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Error message"
        message:
          type: string
          example: "Detailed error description"
    Success:
      type: object
      properties:
        message:
          type: string
          example: "Operation completed successfully"
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Validation failed"
            message: "Required fields are missing"
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Unauthorized"
            message: "Authentication required"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Not found"
            message: "Resource does not exist"
...